Transitioning the installer to rust
im doing this because rust has simply got better I/O and networking support. this may be down to the fact Rust was made in the age of the internet LMFAO anyway yeah im converting it all and its gonna be amazeballs!

i imagine there will still be ways to write parts of the installer in C if a contributor oh-so wanted to but i dont think that is a good idea at all cross compiling on something like this would be awful. maybe on the NTGET Client?